Move to elfinfo and prep to improve
This commit moves work that was done in nightingale back to this project
mostly.  The main difference is that instead of only passing a Elf *
around, we now use a struct elfinfo, which contains pointers to things
that are consistently needed so as to avoid looking up things like the
symbol table repeatedly.

This should be at feature parity with the previous commit as-is, but I
also know there's more work that needs to be done to support things like
the BSS section.  Notes on that are at the bottom of the elf.c file.